# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "Body_ingest_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PDF document to ingest.",
            "title": "File",
            "type": "string"
          },
          "thread_id": {
            "default": "default_session",
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_ingest_document_api_v1_documents_post",
        "type": "object"
      },
...
      "DocumentDeleteResponse": {
        "properties": {
          "cleared": {
            "title": "Cleared",
            "type": "boolean"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "cleared"
        ],
        "title": "DocumentDeleteResponse",
        "type": "object"
      },
      "DocumentIngestResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "document": {
            "title": "Document",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "completed",
              "queued",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "document",
          "thread_id",
          "trace_id",
          "detail"
        ],
        "title": "DocumentIngestResponse",
        "type": "object"
      },
      "DocumentListResponse": {
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "documents"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id",
          "retryable"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nEnterprise Agentic RAG API.\n\n* `POST /api/v1/chat` — buffered answer\n* `POST /api/v1/chat/stream` — SSE stream (`node`, `tool_call`, `observation`, `token`, `done` | `error`)\n* `POST|GET|DELETE /api/v1/documents` — per-session knowledge base\n* `GET /healthz`, `GET /readyz` — probes\n\nAuthenticate with `X-API-Key`. The tenant is derived from the key, never from the\nrequest body, so a caller cannot address another tenant's data.\n",
    "title": "Agentic RAG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents": {
      "delete": {
        "operationId": "clear_documents_api_v1_documents_delete",
        "parameters": [
          {
            "description": "Conversation scope whose knowledge base is addressed.",
            "in": "query",
            "name": "thread_id",
            "required": false,
            "schema": {
              "default": "default_session",
              "description": "Conversation scope whose knowledge base is addressed.",
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant API key.",
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant API key.",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Clear the session knowledge base",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "description": "Conversation scope whose knowledge base is addressed.",
            "in": "query",
            "name": "thread_id",
            "required": false,
            "schema": {
              "default": "default_session",
              "description": "Conversation scope whose knowledge base is addressed.",
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "description": "Tenant API key.",
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant API key.",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List documents in the session knowledge base",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "Ingestion is synchronous in Phase 1 but already reports a job_id and 202, so moving it onto the Phase 5 queue is not a contract change.",
        "operationId": "ingest_document_api_v1_documents_post",
        "parameters": [
          {
            "description": "Tenant API key.",
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tenant API key.",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_document_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ingest a PDF into the session knowledge base",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}